+++ crates/sncast/src/response/structs.rs
@@ -108,15 +108,15 @@ impl CommandResponse for MulticallNewResponse {}
 #[derive(Serialize)]
 pub struct ShowConfigResponse {
     pub profile: Option<String>,
-    pub chain_id: String,
+    pub chain_id: Option<String>,
     pub rpc_url: Option<String>,
     pub account: Option<String>,
     pub accounts_file_path: Option<Utf8PathBuf>,
     pub keystore: Option<Utf8PathBuf>,
     pub wait_timeout: Option<Decimal>,
     pub wait_retry_interval: Option<Decimal>,
-    pub block_explorer: Option<block_explorer::Service>,
     pub show_explorer_links: bool,
+    pub block_explorer: Option<block_explorer::Service>,
 }
 impl CommandResponse for ShowConfigResponse {}