+++ crates/sncast/src/starknet_commands/show_config.rs
@@ -18,12 +18,17 @@ pub struct ShowConfig {
 #[allow(clippy::ptr_arg)]
 pub async fn show_config(
     show: &ShowConfig,
-    provider: &JsonRpcClient<HttpTransport>,
+    provider: &Option<JsonRpcClient<HttpTransport>>,
     cast_config: CastConfig,
     profile: Option<String>,
 ) -> Result<ShowConfigResponse> {
-    let chain_id_field = get_chain_id(provider).await?;
-    let chain_id = chain_id_to_network_name(chain_id_field);
+    let chain_id = if let Some(provider) = provider {
+        let chain_id_field = get_chain_id(provider).await?;
+        Some(chain_id_to_network_name(chain_id_field))
+    } else {
+        None
+    };
+
     let rpc_url = Some(show.rpc.url.clone().unwrap_or(cast_config.url)).filter(|p| !p.is_empty());
     let account = Some(cast_config.account).filter(|p| !p.is_empty());
     let mut accounts_file_path =
@@ -35,7 +40,6 @@ pub async fn show_config(
     let wait_timeout = Some(cast_config.wait_params.get_timeout());
     let wait_retry_interval = Some(cast_config.wait_params.get_retry_interval());
     let block_explorer = cast_config.block_explorer;
-    let show_explorer_links = cast_config.show_explorer_links;
 
     Ok(ShowConfigResponse {
         profile,
@@ -46,7 +50,7 @@ pub async fn show_config(
         keystore,
         wait_timeout: wait_timeout.map(|x| Decimal(u64::from(x))),
         wait_retry_interval: wait_retry_interval.map(|x| Decimal(u64::from(x))),
+        show_explorer_links: cast_config.show_explorer_links,
         block_explorer,
-        show_explorer_links,
     })
 }