+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Changed
 
 - Changed return type of `declare` in Cairo Deployment Scripts so it can handle already declared contracts without failing
+- Allow using `show-config` command without providing rpc url
 
 ## [0.33.0] - 2024-11-04